Issue #2: Configure BIOS and RAID separately
Previously we were configuring BIOS and RAID simultaneously in order to
avoid an unnecessary reboot. This was found to cause problems on a
particular R630 server, causing the configuration to fail and the
Lifecycle controller to enter a reboot loop. The issue was resolved by
resetting the iDRAC configuration to factory defaults and clearing the
job queue. Rather than go through that process again we'll apply BIOS
and RAID configuration separately.