Calculate RAID size_mb parameter correctly
The current calculation of the size_mb parameter is incorrect for RAID
levels using parity disks, such as RAID 5 or RAID 6. It results in a
virtual disk that is much smaller than expected. For example, a RAID 5
would only be as big as the smallest disk.

This commit uses the formula described in issue #4 [1], but adapted to
support all RAID levels listed in python-dracclient [2].

[1] https://github.com/stackhpc/drac/issues/4#issuecomment-409946211
[2] https://opendev.org/openstack/python-dracclient/src/tag/3.1.1/dracclient/resources/raid.py#L25-L34

Co-authored-by: Mark Goddard <mark@stackhpc.com>