Fix evaluation of check_mode
With drac_check_mode defaulting to null, the expression used for
check_mode was evaluated as an empty string, which failed because
check_mode expects a valid boolean.

Pass true as second argument to the default filter to force use of the
default value. Also add bool filter while here.

Closes #8.